@@ -1,118 +1,103 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>net6.0-android;net6.0-ios;net6.0-maccatalyst;net6.0-macos</TargetFrameworks>
-
+		<TargetFrameworks>net7.0-ios;net7.0-maccatalyst;net7.0-android33.0</TargetFrameworks>
 		<SingleProject>true</SingleProject>
 		<OutputType>Exe</OutputType>
+		<!-- Display name -->
+		<ApplicationTitle>ToDo</ApplicationTitle>
+		<!-- App Identifier -->
+		<ApplicationId>com.unoplatform.ToDo</ApplicationId>
+		<ApplicationIdGuid>60E8B07F-8E27-4060-85E8-F0F301846722</ApplicationIdGuid>
+		<!-- Versions -->
+		<ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
+		<ApplicationVersion>1</ApplicationVersion>
 
-		<RuntimeIdentifier Condition="'$(TargetFramework)' == 'net6.0-ios'">iossimulator-x64</RuntimeIdentifier>
-		<RuntimeIdentifier Condition="'$(TargetFramework)' == 'net6.0-maccatalyst'">maccatalyst-x64</RuntimeIdentifier>
-		<RuntimeIdentifier Condition="'$(TargetFramework)' == 'net6.0-macos'">osx-x64</RuntimeIdentifier>
+		<AndroidManifest>Android\AndroidManifest.xml</AndroidManifest>
 
 		<!-- Debugger workaround https://github.com/dotnet/maui-samples/blob/8aa6b8780b12e97b157514c3bdc54bb4a13001cd/HelloMacCatalyst/HelloMacCatalyst.csproj#L7 -->
-		<!-- <MtouchExtraArgs Condition="'$(TargetFramework)' == 'net6.0-maccatalyst'">$(MtouchExtraArgs) -setenv:MONO_THREADS_SUSPEND=preemptive</MtouchExtraArgs> -->
+		<!-- <MtouchExtraArgs Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">$(MtouchExtraArgs) -setenv:MONO_THREADS_SUSPEND=preemptive</MtouchExtraArgs> -->
 		<!-- Required for C# Hot Reload -->
-		<!--<UseInterpreter Condition="'$(Configuration)' == 'Debug' and '$(TargetFramework)' != 'net6.0-maccatalyst'">True</UseInterpreter>-->
+		<UseInterpreter Condition="'$(Configuration)' == 'Debug' and $([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) != 'maccatalyst'">True</UseInterpreter>
 		<IsUnoHead>true</IsUnoHead>
-
-		<SupportedOSPlatformVersion Condition="'$(TargetFramework)' == 'net6.0-ios'">14.2</SupportedOSPlatformVersion>
-		<SupportedOSPlatformVersion Condition="'$(TargetFramework)' == 'net6.0-maccatalyst'">14.0</SupportedOSPlatformVersion>
-		<SupportedOSPlatformVersion Condition="'$(TargetFramework)' == 'net6.0-android'">21.0</SupportedOSPlatformVersion>
-		<SupportedOSPlatformVersion Condition="'$(TargetFramework)'=='net6.0-macos'">10.14</SupportedOSPlatformVersion>
+	</PropertyGroup>
+	<PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|net7.0-android|AnyCPU'">
+	  <AndroidEnableMultiDex>True</AndroidEnableMultiDex>
+	</PropertyGroup>
+	<PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|net7.0-android|AnyCPU'">
+	  <AndroidEnableMultiDex>True</AndroidEnableMultiDex>
+	</PropertyGroup>
+	<PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|net7.0-android33.0|AnyCPU'">
+	  <AndroidEnableMultiDex>True</AndroidEnableMultiDex>
+	</PropertyGroup>
+	<PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|net7.0-android33.0|AnyCPU'">
+	  <AndroidEnableMultiDex>True</AndroidEnableMultiDex>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\ToDo\ToDo.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.0.0"/>
-		<PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.1"/>
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0"/>
-		<PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0"/>
-		<PackageReference Include="Refit" Version="6.3.2" />
-		<PackageReference Include="Refit.HttpClientFactory" Version="6.3.2" />
-
-		<PackageReference Include="Uno.WinUI"  Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.WinUI.RemoteControl"  Version="$(UnoVersion)" Condition="'$(Configuration)'=='Debug'" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging"  Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.Material.WinUI" Version="$(UnoMaterialVersion)" />
-		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="$(UnoToolkitVersion)"/>
-		<PackageReference Include="Uno.Toolkit.WinUI" Version="$(UnoToolkitVersion)"/>
-
-		<PackageReference Include="Uno.Extensions.Authentication" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Configuration" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Core" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Hosting" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Http" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Localization" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Logging.Serilog" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive.Messaging" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" Version="$(UnoExtensionsVersion)" />
+		<PackageReference Include="Uno.Extensions.Core" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Core.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Resizetizer" Version="1.2.0-dev.11" />
+		<PackageReference Include="Uno.WinUI" Version="4.10.0-dev.242" />
+		<PackageReference Include="Uno.Extensions.Configuration" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Http" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Http.Refit" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Logging.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Serialization.Http" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Serialization.Refit" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Material.WinUI" Version="2.7.0-dev.11" />
+		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="3.1.0-dev.10" />
+		<PackageReference Include="Uno.Toolkit.WinUI" Version="3.1.0-dev.10" />
+		<PackageReference Include="Uno.Extensions.Authentication.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Hosting.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Localization.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Navigation.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Reactive.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="7.0.0" />
+		<PackageReference Include="Uno.Extensions.Logging.OSLog" Version="1.6.0-dev.2" />
+		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="4.10.0-dev.242" />
+		<PackageReference Include="Uno.WinUI.RemoteControl" Version="4.10.0-dev.242" Condition="'$(Configuration)'=='Debug'" />
 	</ItemGroup>
 	<Choose>
-
-		<When Condition="'$(TargetFramework)'=='net6.0-android'">
+		<When Condition="$(IsAndroid)">
 			<ItemGroup>
-				<PackageReference Include="Xamarin.AndroidX.Lifecycle.LiveData" Version="2.1.0" />
+				<PackageReference Include="Xamarin.Google.Android.Material" Version="1.8.0" />
 				<PackageReference Include="Uno.UniversalImageLoader" Version="1.9.36" />
 			</ItemGroup>
 			<ItemGroup>
 				<AndroidEnvironment Include="Android/environment.conf" />
 			</ItemGroup>
 		</When>
-
-		<When Condition="'$(TargetFramework)'=='net6.0-ios'">
-			<PropertyGroup Condition="'$(TargetFramework)'=='net6.0-ios'">
+		<When Condition="$(IsIOS)">
+			<PropertyGroup>
 				<MtouchExtraArgs>$(MtouchExtraArgs) --setenv=MONO_GC_PARAMS=soft-heap-limit=512m,nursery-size=64m,evacuation-threshold=66,major=marksweep,concurrent-sweep</MtouchExtraArgs>
-				<!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable --registrar:static</MtouchExtraArgs>
+				<!-- See https://github.com/unoplatform/uno/issues/9430 for more details. -->
+				<MtouchExtraArgs>$(MtouchExtraArgs) --registrar:static</MtouchExtraArgs>
+				<RuntimeIdentifier Condition="'$(RuntimeIdentifier)'==''">iossimulator-x64</RuntimeIdentifier>
 			</PropertyGroup>
-			<ItemGroup>
-				<PackageReference Include="Uno.Extensions.Logging.OSLog" Version="1.4.0" />
-
-				<!-- This is disabled because of https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/3127 -->
-				<PackageReference Include="Microsoft.Identity.Client" Version="4.47.2" ExcludeAssets="all" IncludeAssets="none" />
-			</ItemGroup>
+			<PropertyGroup Condition="'$(Configuration)'=='Release'">
+				<!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
+				<MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
+			</PropertyGroup>
 		</When>
-
-		<When Condition="'$(TargetFramework)'=='net6.0-maccatalyst'">
+		<When Condition="$(IsMacCatalyst)">
 			<PropertyGroup>
 				<!-- Configure the GC -->
 				<MtouchExtraArgs>$(MtouchExtraArgs) --setenv=MONO_GC_PARAMS=soft-heap-limit=512m,nursery-size=64m,evacuation-threshold=66,major=marksweep,concurrent-sweep</MtouchExtraArgs>
 				<!-- Required for unknown crash as of .NET 6 Mobile Preview 13 -->
 				<MtouchExtraArgs>$(MtouchExtraArgs) --registrar:static</MtouchExtraArgs>
-				<!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
 				<!-- Full globalization is required for Uno -->
 				<InvariantGlobalization>false</InvariantGlobalization>
+				<RuntimeIdentifier Condition="'$(RuntimeIdentifier)'==''">maccatalyst-x64</RuntimeIdentifier>
 			</PropertyGroup>
-			<ItemGroup>
-				<PackageReference Include="Uno.Extensions.Logging.OSLog" Version="1.4.0" />
-
-				<!-- This is disabled because of https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/3127 -->
-				<PackageReference Include="Microsoft.Identity.Client" Version="4.47.2" ExcludeAssets="all" IncludeAssets="none" />
-			</ItemGroup>
+			<PropertyGroup Condition="'$(Configuration)'=='Release'">
+				<!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
+				<MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
+			</PropertyGroup>
 		</When>
-
-		<When Condition="'$(TargetFramework)'=='net6.0-macos'">
-			<PropertyGroup></PropertyGroup>
-		</When>
-
 	</Choose>
-	<Import Project="..\ToDo.UI\ToDo.UI.projitems" Label="Shared" />
+	<ItemGroup>
+		<ProjectReference Include="..\ToDo\ToDo.csproj" />
+	</ItemGroup>
+	<Import Project="..\ToDo.Base\base.props" />
 </Project>
