@@ -1,64 +1,41 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<OutputType Condition="'$(Configuration)'=='Release'">WinExe</OutputType>
 		<OutputType Condition="'$(Configuration)'=='Debug'">Exe</OutputType>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 	</PropertyGroup>
-	<ItemGroup Condition="exists('..\ToDo.UWP')">
-		<EmbeddedResource Include="..\ToDo.UWP\Package.appxmanifest" LogicalName="Package.appxmanifest" />
-		<Content Include="..\ToDo.UWP\Assets\StoreLogo.png" Link="Assets\StoreLogo.png" />
-		<Content Include="Assets\Fonts\uno-fluentui-assets.ttf" />
-	</ItemGroup>
 	<ItemGroup>
-		<UpToDateCheckInput Include="..\ToDo.UI\**\*.xaml" />
+		<EmbeddedResource Include="Package.appxmanifest" />
 	</ItemGroup>
+
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="7.0.0" />
-		<PackageReference Include="Microsoft.Identity.Client" Version="4.47.2" />
+		<PackageReference Include="Uno.Extensions.Core" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Core.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Resizetizer" Version="1.2.0-dev.11" />
+		<PackageReference Include="Uno.WinUI.Skia.Gtk" Version="4.10.0-dev.242" />
+		<PackageReference Include="Uno.Extensions.Configuration" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Http" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Http.Refit" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Logging.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Serialization.Http" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Serialization.Refit" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Material.WinUI" Version="2.7.0-dev.11" />
+		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="3.1.0-dev.10" />
+		<PackageReference Include="Uno.Toolkit.WinUI" Version="3.1.0-dev.10" />
+		<PackageReference Include="Uno.Extensions.Authentication.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Hosting.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Localization.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Navigation.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Reactive.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="7.0.0" />
 		<PackageReference Include="SkiaSharp.Views.Uno.WinUI" Version="2.88.3" />
 		<PackageReference Include="SkiaSharp.Skottie" Version="2.88.3" />
-
-		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.0.0"/>
-		<PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.1"/>
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0"/>
-		<PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0"/>
-		<PackageReference Include="Refit" Version="6.3.2" />
-		<PackageReference Include="Refit.HttpClientFactory" Version="6.3.2" />
-
-		<PackageReference Include="Uno.WinUI.MSAL" Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.WinUI.Skia.Gtk"  Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.WinUI.RemoteControl"  Version="$(UnoVersion)" Condition="'$(Configuration)'=='Debug'" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging"  Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.Material.WinUI" Version="$(UnoMaterialVersion)" />
-		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="$(UnoToolkitVersion)"/>
-		<PackageReference Include="Uno.Toolkit.WinUI" Version="$(UnoToolkitVersion)"/>
-
-		<PackageReference Include="Uno.Extensions.Authentication" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Configuration" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Core" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Hosting" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Http" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Localization" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Logging.Serilog" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive.Messaging" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" Version="$(UnoExtensionsVersion)" />
+		<PackageReference Include="Uno.WinUI.RemoteControl" Version="4.10.0-dev.242" Condition="'$(Configuration)'=='Debug'" />
+		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="4.10.0-dev.242" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\ToDo\ToDo.csproj" />
 	</ItemGroup>
-	<Import Project="..\ToDo.UI\ToDo.UI.projitems" Label="Shared" />
+	<Import Project="..\ToDo.Base\base.props" />
 </Project>
