@@ -1,138 +1,57 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<OutputType>WinExe</OutputType>
-		<TargetFramework>net6.0-windows10.0.19041.0</TargetFramework>
-		<TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
-		<RootNamespace>ToDo</RootNamespace>
-		<ApplicationManifest>app.manifest</ApplicationManifest>
+		<TargetFramework>net7.0-windows10.0.19041.0</TargetFramework>
+		<TargetPlatformMinVersion>10.0.18362.0</TargetPlatformMinVersion>
+		<RootNamespace>UnoWinUIQuickStart</RootNamespace>
 		<Platforms>x86;x64;arm64</Platforms>
 		<RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
 		<PublishProfile>win10-$(Platform).pubxml</PublishProfile>
 		<UseWinUI>true</UseWinUI>
-		<DefaultLanguage>en</DefaultLanguage>
 		<EnableMsixTooling>true</EnableMsixTooling>
 	</PropertyGroup>
-
+	<PropertyGroup>
+		<!-- Bundles the WinAppSDK binaries (Uncomment for unpackaged builds) -->
+		<!--<WindowsAppSDKSelfContained>true</WindowsAppSDKSelfContained>-->
+		<!-- This bundles the .NET Core libraries (Uncomment for packaged builds)  -->
+		<!--<SelfContained>true</SelfContained>-->
+	</PropertyGroup>
+	
 	<ItemGroup>
-		<Content Include="Images\LargeTile.scale-100.png" />
-		<Content Include="Images\LargeTile.scale-125.png" />
-		<Content Include="Images\LargeTile.scale-150.png" />
-		<Content Include="Images\LargeTile.scale-200.png" />
-		<Content Include="Images\LargeTile.scale-400.png" />
-		<Content Include="Images\LockScreenLogo.scale-200.png" />
-		<Content Include="Images\SmallTile.scale-100.png" />
-		<Content Include="Images\SmallTile.scale-125.png" />
-		<Content Include="Images\SmallTile.scale-150.png" />
-		<Content Include="Images\SmallTile.scale-200.png" />
-		<Content Include="Images\SmallTile.scale-400.png" />
-		<Content Include="Images\SplashScreen.scale-100.png" />
-		<Content Include="Images\SplashScreen.scale-125.png" />
-		<Content Include="Images\SplashScreen.scale-150.png" />
-		<Content Include="Images\SplashScreen.scale-200.png" />
-		<Content Include="Images\SplashScreen.scale-400.png" />
-		<Content Include="Images\Square150x150Logo.scale-100.png" />
-		<Content Include="Images\Square150x150Logo.scale-125.png" />
-		<Content Include="Images\Square150x150Logo.scale-150.png" />
-		<Content Include="Images\Square150x150Logo.scale-200.png" />
-		<Content Include="Images\Square150x150Logo.scale-400.png" />
-		<Content Include="Images\Square44x44Logo.scale-100.png" />
-		<Content Include="Images\Square44x44Logo.scale-125.png" />
-		<Content Include="Images\Square44x44Logo.scale-150.png" />
-		<Content Include="Images\Square44x44Logo.scale-200.png" />
-		<Content Include="Images\Square44x44Logo.scale-400.png" />
-		<Content Include="Images\Square44x44Logo.targetsize-16.png" />
-		<Content Include="Images\Square44x44Logo.targetsize-24.png" />
-		<Content Include="Images\Square44x44Logo.targetsize-24_altform-unplated.png" />
-		<Content Include="Images\Square44x44Logo.targetsize-256.png" />
-		<Content Include="Images\Square44x44Logo.targetsize-32.png" />
-		<Content Include="Images\Square44x44Logo.targetsize-48.png" />
-		<Content Include="Images\StoreLogo.scale-100.png" />
-		<Content Include="Images\StoreLogo.scale-125.png" />
-		<Content Include="Images\StoreLogo.scale-150.png" />
-		<Content Include="Images\StoreLogo.scale-200.png" />
-		<Content Include="Images\StoreLogo.scale-400.png" />
-		<Content Include="Images\Wide310x150Logo.scale-100.png" />
-		<Content Include="Images\Wide310x150Logo.scale-125.png" />
-		<Content Include="Images\Wide310x150Logo.scale-150.png" />
-		<Content Include="Images\Wide310x150Logo.scale-200.png" />
-		<Content Include="Images\Wide310x150Logo.scale-400.png" />
-	</ItemGroup>
-
-	<ItemGroup>
+		<Content Include="Images\**" />
 		<Manifest Include="$(ApplicationManifest)" />
 	</ItemGroup>
+
 	<ItemGroup>
-		<FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" RuntimeFrameworkVersion="10.0.22000.25" />
-		<FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" TargetingPackVersion="10.0.22000.25" />
-	</ItemGroup>
-	<ItemGroup>
-		<PackageReference Include="Microsoft.Identity.Client" Version="4.47.2" />
-		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.1.5" />
-		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.755" />
+		<PackageReference Include="Uno.Extensions.Core" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Core.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Resizetizer" Version="1.2.0-dev.11" />
+		<PackageReference Include="Uno.WinUI" Version="4.10.0-dev.242" />
+		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.3.230502000" />
+		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.756" />
+		<PackageReference Include="Uno.Extensions.Configuration" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Http" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Http.Refit" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Logging.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Serialization.Http" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Serialization.Refit" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Material.WinUI" Version="2.7.0-dev.11" />
+		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="3.1.0-dev.10" />
+		<PackageReference Include="Uno.Toolkit.WinUI" Version="3.1.0-dev.10" />
+		<PackageReference Include="Uno.Extensions.Authentication.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Hosting.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Localization.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Navigation.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Uno.Extensions.Reactive.WinUI" Version="2.5.0-dev.66" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="7.0.0" />
 		<PackageReference Include="Uno.Core.Extensions.Logging.Singleton" Version="4.0.1" />
-		<PackageReference Include="System.Net.Http.WinHttpHandler" Version="6.0.1" />
-
-		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.0.0"/>
-		<PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.1"/>
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0"/>
-		<PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0"/>
-		<PackageReference Include="Refit" Version="6.3.2" />
-		<PackageReference Include="Refit.HttpClientFactory" Version="6.3.2" />
-
-		<PackageReference Include="Uno.WinUI.MSAL" Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.WinUI.Skia.Wpf"  Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.WinUI.RemoteControl"  Version="$(UnoVersion)" Condition="'$(Configuration)'=='Debug'" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging"  Version="$(UnoVersion)" />
-		<PackageReference Include="Uno.Material.WinUI" Version="$(UnoMaterialVersion)" />
-		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="$(UnoToolkitVersion)"/>
-		<PackageReference Include="Uno.Toolkit.WinUI" Version="$(UnoToolkitVersion)"/>
-
-		<PackageReference Include="Uno.Extensions.Authentication" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Configuration" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Core" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Hosting" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Http" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Localization" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Logging.Serilog" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive.Messaging" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Reactive.WinUI" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" Version="$(UnoExtensionsVersion)" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" Version="$(UnoExtensionsVersion)" />
+		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="4.10.0-dev.242" />
 	</ItemGroup>
+	
 	<ItemGroup>
 		<ProjectReference Include="..\ToDo\ToDo.csproj" />
 	</ItemGroup>
 
-	<!-- 
-    Defining the "Msix" ProjectCapability here allows the Single-project MSIX Packaging
-    Tools extension to be activated for this project even if the Windows App SDK Nuget
-    package has not yet been restored.
-  -->
-	<ItemGroup Condition="'$(DisableMsixProjectCapabilityAddedByProject)'!='true' and '$(EnableMsixTooling)'=='true'">
-		<ProjectCapability Include="Msix" />
-	</ItemGroup>
-
-	<!-- 
-    Defining the "HasPackageAndPublishMenuAddedByProject" property here allows the Solution 
-    Explorer "Package and Publish" context menu entry to be enabled for this project even if 
-    the Windows App SDK Nuget package has not yet been restored.
-  -->
-	<PropertyGroup Condition="'$(DisableHasPackageAndPublishMenuAddedByProject)'!='true' and '$(EnableMsixTooling)'=='true'">
-		<HasPackageAndPublishMenu>true</HasPackageAndPublishMenu>
-	</PropertyGroup>
-
-	<Import Project="..\ToDo.UI\ToDo.UI.projitems" Label="Shared" />
-
+	<Import Project="..\ToDo.Base\base.props" />
 </Project>
