@@ -2,21 +2,20 @@ using System;
 using GLib;
 using Uno.UI.Runtime.Skia;
 
-namespace ToDo.Skia.Gtk
+namespace ToDo.Skia.Gtk;
+
+public class Program
 {
-	class Program
+	public static void Main(string[] args)
 	{
-		static void Main(string[] args)
+		ExceptionManager.UnhandledException += delegate (UnhandledExceptionArgs expArgs)
 		{
-			ExceptionManager.UnhandledException += delegate (UnhandledExceptionArgs expArgs)
-			{
-				Console.WriteLine("GLIB UNHANDLED EXCEPTION" + expArgs.ExceptionObject.ToString());
-				expArgs.ExitApplication = true;
-			};
+			Console.WriteLine("GLIB UNHANDLED EXCEPTION" + expArgs.ExceptionObject.ToString());
+			expArgs.ExitApplication = true;
+		};
 
-			var host = new GtkHost(() => new App(), args);
+		var host = new GtkHost(() => new AppHead(), args);
 
-			host.Run();
-		}
+		host.Run();
 	}
 }
