@@ -1,15 +1,15 @@
-﻿using UIKit;
+using UIKit;
 
-namespace ToDo
+namespace ToDo.MacCatalyst
 {
-    public class EntryPoint
-    {
-        // This is the main entry point of the application.
-        static void Main(string[] args)
-        {
-            // if you want to use a different Application Delegate class from "AppDelegate"
-            // you can specify it here.
-            UIApplication.Main(args, null, typeof(App));
-        }
-    }
+	public class EntryPoint
+	{
+		// This is the main entry point of the application.
+		public static void Main(string[] args)
+		{
+			// if you want to use a different Application Delegate class from "AppDelegate"
+			// you can specify it here.
+			UIApplication.Main(args, null, typeof(AppHead));
+		}
+	}
 }
