@@ -26,19 +26,19 @@ public class MockTaskEndpoint : ITaskEndpoint
 	{
 		var tasks = await _listEndpoint.LoadListTasks(todoTaskListId);
 
-		return new TaskReponseData<TaskData> { Value = tasks.ToList() };
+		return new TaskReponseData<TaskData> { Value = tasks?.ToList() };
 	}
 
 	public async Task<TaskReponseData<TaskData>> GetAllAsync(CancellationToken ct) => await _listEndpoint.GetAllTasksAsync( ct: ct);
 
-	public async Task<TaskData> GetAsync(string listId, string taskId, CancellationToken ct)
+	public async Task<TaskData?> GetAsync(string listId, string taskId, CancellationToken ct)
 	{
 		var tasks = await GetAsync(listId, ct);
 
-		return tasks.Value.First(x => x.Id == taskId);
+		return tasks?.Value?.First(x => x.Id == taskId);
 	}
 
-	public async Task<TaskReponseData<TaskData>> GetByFilterAsync(string displayName, CancellationToken ct) => await _listEndpoint.GetAllTasksAsync(displayName, ct);
+	public async Task<TaskReponseData<TaskData>?> GetByFilterAsync(string displayName, CancellationToken ct) => await _listEndpoint.GetAllTasksAsync(displayName, ct);
 
 	public async Task<TaskData> UpdateAsync(string listId, string taskId, [Body] TaskData updatedTask, CancellationToken ct)
 	{
