@@ -24,16 +24,17 @@ public class MockTaskListEndpoint : ITaskListEndpoint
 	private IList<TaskData>? allTasks;
 	private IDictionary<string, IList<TaskData>> taskData = new Dictionary<string, IList<TaskData>>();
 
-	private async Task<IList<TaskListData>> Load()
+	private async Task<IList<TaskListData>?> Load()
 	{
 		if (data is null)
 		{
-			data = (await _dataService.ReadPackageFileAsync<TaskListData[]>(_listSerializer, ListDataFile)).ToList();
+			data = await _dataService.ReadPackageFileAsync<TaskListData[]>(_listSerializer, ListDataFile);
 		}
-		return data;
+
+		return data?.ToList();
 	}
 
-	internal async Task<IList<TaskData>> LoadListTasks(string listId)
+	internal async Task<IList<TaskData>?> LoadListTasks(string listId)
 	{
 		if (taskData.TryGetValue(listId, out var existingTasks))
 		{
@@ -42,17 +43,22 @@ public class MockTaskListEndpoint : ITaskListEndpoint
 
 		_ = await LoadAllTasks();
 
-		taskData[listId] = allTasks.Where(x=>x.ParentList?.Id==listId).ToList();
+		if (allTasks is not null)
+		{
+			taskData[listId] = allTasks.Where(x => x.ParentList?.Id == listId).ToList();
+		}
+
 		return taskData[listId];
 	}
 
-	private async Task<IList<TaskData>> LoadAllTasks()
+	private async Task<IList<TaskData>?> LoadAllTasks()
 	{
 		if (allTasks is null)
 		{
-			allTasks = (await _dataService.ReadPackageFileAsync<TaskData[]>(_taskSerializer, TasksDataFile)).ToList();
+			allTasks = await _dataService.ReadPackageFileAsync<TaskData[]>(_taskSerializer, TasksDataFile);
 		}
-		return allTasks;
+
+		return allTasks?.ToList();
 	}
 
 	public async Task<TaskListData> CreateAsync(TaskListRequestData todoList, CancellationToken ct)
@@ -67,27 +73,39 @@ public class MockTaskListEndpoint : ITaskListEndpoint
 	public async Task<HttpResponseMessage> DeleteAsync(string todoTaskListId, CancellationToken ct)
 	{
 		_ = await Load();
-		var list = data.FirstOrDefault(x => x.Id == todoTaskListId);
-		if (list is not null)
+
+		if (data is not null)
 		{
-			data?.Remove(list);
+			var list = data.FirstOrDefault(x => x.Id == todoTaskListId);
+
+			if (list is not null)
+			{
+				data?.Remove(list);
+			}
 		}
+
 		return new HttpResponseMessage(System.Net.HttpStatusCode.OK);
 	}
 
 	public async Task<TaskReponseData<TaskListData>> GetAllAsync(CancellationToken ct)
 	{
 		_ = await Load();
-		return new TaskReponseData<TaskListData> { Value = data.ToList() };
+
+		return new TaskReponseData<TaskListData> { Value = data?.ToList() };
 	}
 
 	public async Task<TaskListData> GetAsync(string todoTaskListId, CancellationToken ct)
 	{
 		_ = await Load();
-		var list = data.FirstOrDefault(x => x.Id == todoTaskListId);
-		if (list is not null)
+
+		if (data is not null)
 		{
-			return list;
+			var list = data.FirstOrDefault(x => x.Id == todoTaskListId);
+
+			if (list is not null)
+			{
+				return list;
+			}
 		}
 
 		return new TaskListData();
@@ -97,11 +115,14 @@ public class MockTaskListEndpoint : ITaskListEndpoint
 	{
 		_ = await Load();
 
-		var list = data.FirstOrDefault(x => x.Id == todoTaskListId);
-		if (list is not null)
+		if (data is not null)
 		{
-			list.DisplayName = todoList.DisplayName;
-			return list;
+			var list = data.FirstOrDefault(x => x.Id == todoTaskListId);
+			if (list is not null)
+			{
+				list.DisplayName = todoList.DisplayName;
+				return list;
+			}
 		}
 
 		return new TaskListData();
@@ -111,16 +132,25 @@ public class MockTaskListEndpoint : ITaskListEndpoint
 	internal async Task AddTaskToList(string todoTaskListId, TaskData task)
 	{
 		var list = await LoadListTasks(todoTaskListId);
-		list.Add(task);
+
+		if (list is not null)
+		{
+			list.Add(task);
+		}
+
 	}
 
 	internal async Task DeleteTaskFromList(string todoTaskListId, string taskId)
 	{
 		var list = await LoadListTasks(todoTaskListId);
-		var task = list.FirstOrDefault(t => t.Id == taskId);
-		if (task is not null)
+
+		if (list is not null)
 		{
-			list.Remove(task);
+			var task = list.FirstOrDefault(t => t.Id == taskId);
+			if (task is not null)
+			{
+				list.Remove(task);
+			}
 		}
 	}
 
@@ -134,9 +164,11 @@ public class MockTaskListEndpoint : ITaskListEndpoint
 	{
 		var tasks = await LoadAllTasks();
 
-		if(string.IsNullOrWhiteSpace(displayName))
-			return new TaskReponseData<TaskData> { Value = tasks.ToList() };
+		if (string.IsNullOrWhiteSpace(displayName))
+		{
+			return new TaskReponseData<TaskData> { Value = tasks?.ToList() };
+		}
 
-		return new TaskReponseData<TaskData> { Value = tasks.Where(x => x.Title != null && x.Title.Contains(displayName)).ToList() };
+		return new TaskReponseData<TaskData> { Value = tasks?.Where(x => x.Title != null && x.Title.Contains(displayName)).ToList() };
 	}
 }
