@@ -1,43 +1,69 @@
 ﻿global using System;
 global using System.Collections.Generic;
 global using System.Collections.Immutable;
-global using System.Diagnostics.Contracts;
-global using System.Globalization;
 global using System.IdentityModel.Tokens.Jwt;
 global using System.Linq;
+global using System.Net.Http;
+global using System.Threading;
+global using System.Threading.Tasks;
+global using System.Collections.Specialized;
+global using System.Text;
+global using System.Diagnostics.Contracts;
+global using System.Globalization;
 global using System.Runtime.CompilerServices;
 global using System.Text.Json;
 global using System.Text.Json.Serialization;
-global using System.Threading;
-global using System.Threading.Tasks;
-global using System.Windows.Input;
-global using System.Net.Http;
 global using System.IO;
-global using CommunityToolkit.Mvvm.Input;
 global using CommunityToolkit.Mvvm.Messaging;
+global using Refit;
+global using ToDo.Business;
+global using ToDo.Business.Mock;
+global using ToDo.Business.Services;
+global using ToDo.Configuration;
+global using ToDo.Data;
+global using ToDo.Data.Models;
+global using ToDo.Presentation.Dialogs;
+global using ToDo.Business.Models;
+global using ToDo.Infrastructure;
+global using ToDo.Presentation;
+global using ToDo.Views;
+global using ToDo.Views.Dialogs;
+global using ToDo.Data.Mock;
+global using Uno.Extensions;
+global using Uno.Extensions.Configuration;
+global using Uno.Extensions.Hosting;
+global using Uno.Extensions.Http;
+global using Uno.Extensions.Localization;
+global using Uno.Extensions.Logging;
+global using Uno.Extensions.Navigation;
+global using Uno.Extensions.Reactive.Messaging;
+global using Uno.Extensions.Navigation.Toolkit;
+global using Uno.Extensions.Navigation.Regions;
+global using Uno.Extensions.Reactive;
+global using Uno.Extensions.Serialization;
+global using Uno.Extensions.Storage;
+global using Uno.UI.MSAL;
+global using Uno.Toolkit.UI;
+global using Windows.ApplicationModel;
+global using Windows.ApplicationModel.Core;
+global using Windows.Foundation;
+global using Windows.Security.Authentication.Web;
+global using Windows.UI.Popups;
+global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Localization;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Options;
 global using Microsoft.Identity.Client;
-global using Refit;
-global using ToDo.Business;
-global using ToDo.Business.Mock;
-global using ToDo.Business.Models;
-global using ToDo.Business.Services;
-global using ToDo.Configuration;
-global using ToDo.Data;
-global using ToDo.Data.Models;
-global using ToDo.Presentation.Dialogs;
-global using Uno.Extensions;
-global using Uno.Extensions.Configuration;
-global using Uno.Extensions.Http;
-global using Uno.Extensions.Localization;
-global using Uno.Extensions.Navigation;
-global using Uno.Extensions.Reactive;
-global using Uno.Extensions.Reactive.Messaging;
-global using Uno.Extensions.Serialization;
-global using Uno.Extensions.Storage;
-global using Uno.UI.MSAL;
-global using Windows.Security.Authentication.Web;
+global using Microsoft.UI.Xaml;
+global using Microsoft.UI.Xaml.Controls;
+global using Microsoft.UI.Xaml.Media;
+global using Microsoft.UI.Xaml.Navigation;
+global using Microsoft.UI.Dispatching;
+global using Microsoft.UI.Xaml.Input;
+global using Microsoft.UI.Xaml.Controls.Primitives;
+global using Microsoft.UI.Xaml.Markup;
+global using Microsoft.UI.Xaml.Data;
+global using Application = Microsoft.UI.Xaml.Application;
+global using ApplicationExecutionState = Windows.ApplicationModel.Activation.ApplicationExecutionState;
